10_lesson/requirements.txt
selenium
requests
pytest
webdriver-manager
allure-pytest 